fixing warning refs #12
git-svn-id: file:///home/behr_mi/git/sumo_synched/trunk@11177 afbd958f-9f77-42d5-a016-97a22340ccf4

+++ sumo/src/netbuild/NBTrafficLightLogic.cpp
@@ -95,7 +95,7 @@ NBTrafficLightLogic::addStep(SUMOTime duration, const std::string &state, int in
         throw ProcessError("When adding phase: illegal character '" + toString(state[illegal]) + "' in state");
     }
     // interpret index
-    if (index < 0 || index >= myPhases.size()) {
+    if (index < 0 || index >= (int)myPhases.size()) {
         // insert at the end
         index = myPhases.size();
     }
